requirements.txt
asgiref==3.5.1
backports.zoneinfo==0.2.1
certifi==2021.10.8
cfgv==3.3.1
charset-normalizer==2.0.12
distlib==0.3.4
Django==4.0.4
filelock==3.6.0
gunicorn==20.1.0
identify==2.5.0
idna==3.3
mailerlite-api-python==0.10.0
mailersend==0.2.1
nodeenv==1.6.0
platformdirs==2.5.2
pre-commit==2.19.0
psycopg2-binary==2.9.3
PyYAML==6.0
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
urllib3==1.26.9
whitenoise==6.1.0
